fix(shared): remove circular dependencies
